// webapp/src/features/range-config/types/range-config.ts
// Copyright 2023 Two Six Technologies
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//

export type NodeConfig = {
  name: string;
  type: 'client' | 'server';
  enclave: string;
  nat: boolean;
  identities: string[];
  genesis: boolean;
  gpu: boolean;
  bridge: boolean;
  platform: 'linux' | 'android';
  architecture: 'x86_64' | 'arm64-v8a' | 'auto';
  environment: string;
};

export type PortMapping = {
  hosts: string[];
  port: string;
};

export type EnclaveConfig = {
  name: string;
  ip: string;
  hosts: string[];
  port_mapping: Record<string, PortMapping>;
};

export type AccessConfig = {
  protocol: string;
  url: string;
  userFmt: string;
  password: string;
};

export type ServiceConfig = {
  name: string;
  access: AccessConfig[];
  type: string;
  'auth-req-view': string;
  'auth-req-reply': string;
  'auth-req-post': string;
  'auth-req_delete': string;
};

export type RangeConfig = {
  range: {
    name: string;
    bastion: {
      range_ip: string;
    };
    RACE_nodes: NodeConfig[];
    enclaves: EnclaveConfig[];
    services: ServiceConfig[];
  };
};